объединил модули. удалил обработчики событий
