+++ js/form.js
@@ -1,7 +1,6 @@
 import { isEcapeKey } from './util.js';
-import { initScale } from './scale.js';
-import { initFilter } from './filter.js';
-import { initSlider } from './slider.js';
+import { pristine } from './validate.js';
+import { initPopup, destroyPopup } from './popup.js';
 
 const bodyElement = document.querySelector('body');
 const formElement = document.querySelector('.img-upload__form');
@@ -9,16 +8,11 @@ const formOverlayElement = formElement.querySelector('.img-upload__overlay');
 const btnCancel = formElement.querySelector('.img-upload__cancel');
 const btnUploadFile = formElement.querySelector('#upload-file');
 
-const pristine = new Pristine(formElement, {
-  classTo: 'img-upload__text',
-  errorTextParent: 'img-upload__text',
-  errorTextClass: 'img-upload__text__error-text'
-});
-
 const openModal = () => {
   formOverlayElement.classList.remove('hidden');
   bodyElement.classList.add('modal-open');
   document.addEventListener('keydown', onEscKeydown);
+  initPopup();
 };
 
 const closeModal = () => {
@@ -27,10 +21,11 @@ const closeModal = () => {
   formOverlayElement.classList.add('hidden');
   bodyElement.classList.remove('modal-open');
   document.removeEventListener('keydown', onEscKeydown);
+  destroyPopup();
 };
 
-const onFileInputChange = () => openModal();
-const onCancel = () => closeModal();
+const onBtnUploadFileChenge = () => openModal();
+const onClickBtnCancel = () => closeModal();
 
 const onFormSubmit = (evt) => {
   evt.preventDefault();
@@ -38,20 +33,16 @@ const onFormSubmit = (evt) => {
 };
 
 const initModal = () => {
-  btnUploadFile.addEventListener('change', onFileInputChange);
-  btnCancel.addEventListener('click', onCancel);
+  btnUploadFile.addEventListener('change', onBtnUploadFileChenge);
+  btnCancel.addEventListener('click', onClickBtnCancel);
   formElement.addEventListener('submit', onFormSubmit);
 };
 
 function onEscKeydown(evt) {
   if (isEcapeKey(evt)) {
     evt.preventDefault();
-    onCancel();
+    onClickBtnCancel();
   }
 }
-
-initScale();
-initFilter();
-initSlider();
 
 export { initModal };
